Speculative crash fix. Various random crashes have been reported, with TaskCanceledException from HttpClient methods being the culprit, of those crashes fully reported. It is suspected, based on the reported behavioral observations of the other crashes, that the underlying fault of those crashes is in fact the same thing.
- The Suspected Problem -
The Imgur.API library uses the System.Net.Http.HttpClient interface for interacting with Imgur's web API. Exceptions thrown from method calls to this interface are not handled, and are propagated up to the caller of the library.
Contrary to documented behavior, methods in HttpClient will throw a TaskCanceledException if the period specified in the Timeout property elapses before the method completes. Despite what the property name suggests, this is not a network timeout; the timeout starts from the point at which the HttpClient call is made, after which the call will be queued if there is no connection available in the pool of network connections that HttpClient maintains. It is therefore possible for the timeout to expire before the HttpClient method call exits the queue of pending calls to complete, upon which TaskCanceledException is thrown.
The effects of this problem are that, under testing conditions, things will proceed normally. Whereas under heavy loads, if many requests are issued to HttpClient in parallel, the timeout for certain requests may expire before the pooled connection resources managed by HttpClient become sufficiently available to service and complete the request (Although how a pending HttpClient request could remain in the queue for the entire default 100 second timeout period is currently unknown).

- Ideal Solution: Disable Method Timeout behavior -
The HttpClient.Timeout property controls the timeout behavior of HttpClient methods, causing it to throw TaskCanceledException. Ideally this behavior would be disabled by setting the property to infinite, but the property is not accessible through the Imgur.API library. Additionally, this may be unstable; part of the undocumented behavior of HttpClient may be such that lower-level timeouts are disabled in favor of the HttpClient timeout. Waiting indefinitely for IO operations is not stable, so keeping these lower-level protocol/network timeouts is desirable. Granularity of timeouts surrounding HttpClient may be a feature yet to emerge in .NET.

- Implemented Solution: Eliminate Parallelism -
Since too many parallel requests to HttpClient may cause problems, all Imgur calls during an operation can be serialized. Previously, a degree of parallelism was implemented when processing Comments from Taggers, to help speed the process up. This parallelism in the Comment processing activity of the application has been eliminated; within this activity, there will be no more than a single pending call to the Imgur API, and thus HttpClient. The drawback to this solution is that the time taken to process Comments may take longer, since parallelism in the activity has been eliminated.
Note that multiple pending requests in HttpClient are still possible application-wide, such as if a user executes a Discord command while the application is processing Comments.

- Alternative Solution: Increase Capacity for Parallel Requests -
Since HttpClient has problems dealing with many parallel requests with its default connection pool, the connection pool can be configured to be larger, so that many parallel requests can be serviced without being queued. This can be configured using the System.Net.ServicePointManager interface, specifically the DefaultConnectionLimit property. Some experimentation would be needed to determine an adequate pool size, which would ideally be configurable in the application's Settings.xml, as different deployment environments of the application may have different constraints.

Actual commit changes:
Removed parallelism in Comment processing, to speculatively fix reported crashes
Potential TaskCanceledException exceptions from Imgur.API library calls caught, and treated as generic interface errors
Application now outputs timestamp of any crash